Add getDeck and getFlashcard middleware, refactor jwtAuth middleware, remove confirmUser middleware

+++ routes/decks.js
@@ -2,18 +2,22 @@ const express = require('express')
 const router = express.Router()
 const User = require('../models/user')
 const jwtAuth = require('../middleware/jwtAuth');
-const confirmUser = require('../middleware/confirmUser')
 
 /* TODO:
     - Update name of deck
     - Get all decks given user id
 */
 
-// get all for current authed user
-router.get('/', jwtAuth, confirmUser, async (req, res) => {
+// get all
+router.get('/', jwtAuth, async (req, res) => {
     try {
+        const user = await User.findById(req.user.id)
 
-        const decks = req.user.library.map(deck => ({
+        if (!user){
+            res.status(404).json({ message: "User not found" })
+        }
+
+        const decks = user.library.map(deck => ({
             id: deck.id,
             name: deck.name
         }))
@@ -26,10 +30,14 @@ router.get('/', jwtAuth, confirmUser, async (req, res) => {
 
 
 // get one from current user's library
-router.get('/:id', jwtAuth, confirmUser, async (req, res) => {
+router.get('/:id', jwtAuth, async (req, res) => {
     try {
+        const user = await User.findById(req.user.id)
+        if (!user) {
+            res.status(404).json({ message: "User not found" })
+        }
 
-        const deck = req.user.library.id(req.params.id)
+        const deck = user.library.id(req.params.id)
         if (!deck) {
             res.status(404).json({ message: "Deck not found" })
         }
@@ -48,18 +56,22 @@ router.get('/:id', jwtAuth, confirmUser, async (req, res) => {
 })
 
 // create one for current user
-router.post('/', jwtAuth, confirmUser, async (req, res) => {
+router.post('/', jwtAuth, async (req, res) => {
     try{
         const deckName = req.body.name;
-    
+
+        const user = await User.findById(req.user.id)
+        if (!user) {
+            res.status(404).json({ message: "User not found"})
+        }        
 
         const newDeck = {
             name: deckName,
             flashcards: []
         }
 
-        req.user.library.push(newDeck);
-        await req.user.save();
+        user.library.push(newDeck);
+        await user.save();
 
         res.status(200).json({ message: "Deck created"})
 
@@ -70,16 +82,20 @@ router.post('/', jwtAuth, confirmUser, async (req, res) => {
 
 
 // delete one of the current user's decks
-router.delete('/:id', jwtAuth, confirmUser, async (req, res) =>{
+router.delete('/:id', jwtAuth, async (req, res) =>{
     try {
+        const user = await User.findById(req.user.id)
+        if (!user) {
+            return res.status(404).json({ message: "User not found" })
+        }
 
-        const deck = req.user.library.id(req.params.id)
+        const deck = user.library.id(req.params.id)
         if (!deck) {
             return res.status(404).json({ message: "Deck not found" })
         }
 
         deck.deleteOne()
-        await req.user.save();
+        await user.save();
 
         return res.status(200).json({ message: "Deck deleted" })
